Refactor earthquakes rendering so it can be reused by volcanoes later
[#162386261]
